requirements.txt
python-telegram-bot>=20.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-genai>=0.1.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
markdown2>=2.4.0
psutil>=5.9.0
